Short method exclusion logic implementation
Extended the instrumenter to exclude methods with less instruction than a certain limit.
This exclusion is opt-in, be default no method will be excluded based on length.
Added a few tests suits for covering the implementation. Sadly invalid compiler argument cannot be tested as it would break the actual test compilation.
Documentation is also updated with this new logic, but will need further adjustments as env. variables are no longer supported yet the documentation still references them.

Fixes #46